git: move GitSettings out of settings module
This type wouldn't be needed if the "git" feature is disabled.
