+++ lib/tests/test_git.rs
@@ -49,6 +49,7 @@ use jj_lib::git::GitPushStats;
 use jj_lib::git::GitRefKind;
 use jj_lib::git::GitRefUpdate;
 use jj_lib::git::GitResetHeadError;
+use jj_lib::git::GitSettings;
 use jj_lib::git::IgnoredRefspec;
 use jj_lib::git::IgnoredRefspecs;
 use jj_lib::git::expand_default_fetch_refspecs;
@@ -71,7 +72,6 @@ use jj_lib::refs::BookmarkPushUpdate;
 use jj_lib::repo::MutableRepo;
 use jj_lib::repo::ReadonlyRepo;
 use jj_lib::repo::Repo as _;
-use jj_lib::settings::GitSettings;
 use jj_lib::settings::RemoteSettings;
 use jj_lib::settings::UserSettings;
 use jj_lib::signing::Signer;
